+++ src/components/Report/PdfReports/ReportContainer.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next';
 import { FC, memo } from 'react';
 import { IBasicReportData, ReportType, Reports } from '@/interfaces/reportInterfaces';
 import { IProject } from '@/interfaces/projectInterfaces';
-import DocumentHeader from './DocumentHeader';
+import DocumentHeader from './reportHeaders/DocumentHeader';
 import ReportTable from './ReportTable';
 import StrategyReportFooter from './StrategyReportFooter';
 import DefaultReportFooter from './DefaultReportFooter';
